Isolate execution polling response projection
